refactor: Abstracted biome-related serialization for recipe code.
