+++ src/main/java/com/gildedgames/aether/event/hooks/DimensionHooks.java
@@ -3,9 +3,9 @@ package com.gildedgames.aether.event.hooks;
 import com.gildedgames.aether.block.AetherBlocks;
 import com.gildedgames.aether.AetherTags;
 import com.gildedgames.aether.recipe.AetherRecipeTypes;
-import com.gildedgames.aether.recipe.recipes.BlockBanRecipe;
-import com.gildedgames.aether.recipe.recipes.ItemBanRecipe;
-import com.gildedgames.aether.recipe.recipes.PlacementConversionRecipe;
+import com.gildedgames.aether.recipe.recipes.ban.BlockBanRecipe;
+import com.gildedgames.aether.recipe.recipes.ban.ItemBanRecipe;
+import com.gildedgames.aether.recipe.recipes.block.PlacementConversionRecipe;
 import com.gildedgames.aether.util.LevelUtil;
 import com.gildedgames.aether.data.resources.AetherDimensions;
 import com.gildedgames.aether.world.AetherTeleporter;
