+++ src/main/java/com/gildedgames/aether/recipe/builder/BlockBanBuilder.java
@@ -1,7 +1,7 @@
 package com.gildedgames.aether.recipe.builder;
 
 import com.gildedgames.aether.recipe.BlockStateIngredient;
-import com.gildedgames.aether.recipe.recipes.BlockBanRecipe;
+import com.gildedgames.aether.recipe.recipes.ban.BlockBanRecipe;
 import com.gildedgames.aether.recipe.serializer.PlacementBanRecipeSerializer;
 import com.google.gson.JsonObject;
 import net.minecraft.data.recipes.FinishedRecipe;
