+++ src/main/java/com/gildedgames/aether/recipe/builder/IncubationBuilder.java
@@ -1,6 +1,6 @@
 package com.gildedgames.aether.recipe.builder;
 
-import com.gildedgames.aether.recipe.recipes.IncubationRecipe;
+import com.gildedgames.aether.recipe.recipes.item.IncubationRecipe;
 import com.google.gson.JsonObject;
 import net.minecraft.advancements.Advancement;
 import net.minecraft.advancements.AdvancementRewards;
