+++ src/main/java/com/gildedgames/aether/recipe/serializer/BlockStateRecipeSerializer.java
@@ -2,7 +2,7 @@ package com.gildedgames.aether.recipe.serializer;
 
 import com.gildedgames.aether.recipe.BlockPropertyPair;
 import com.gildedgames.aether.recipe.BlockStateIngredient;
-import com.gildedgames.aether.recipe.recipes.AbstractBlockStateRecipe;
+import com.gildedgames.aether.recipe.recipes.block.AbstractBlockStateRecipe;
 import com.gildedgames.aether.util.BlockStateRecipeUtil;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
